signal and heredoc fixes
