--- o
@@ -1,11 +1,0 @@
-Makefile
-README.md
-bin
-hi
-include
-l
-lib
-minishell
-o
-src
-wtf
