+++ src/executor/signals.c
@@ -6,7 +6,7 @@
 /*   By: ekordi <ekordi@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/20 14:06:11 by ekordi            #+#    #+#             */
-/*   Updated: 2024/01/20 14:06:13 by ekordi           ###   ########.fr       */
+/*   Updated: 2024/01/20 18:20:15 by ekordi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,4 +58,3 @@ void	setup_signals(void)
 	sigaction(SIGINT, &sa_int, NULL);
 	signal(SIGQUIT, SIG_IGN);
 }
-
